refactor tests
- move all tests to __tests__ folder
- rename `.test.js` to `.spec.js`
- mock db in jest setup
