+++ app/components/ActivateContract/__tests__/ActivateContract.spec.js
@@ -1,10 +1,9 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import ContractState from '../../states/contract-state'
-import BalancesState from '../../states/balances-state'
-
-import ActivateContractContainer, { calcMaxBlocksForContract } from './ActivateContract'
+import ContractState from '../../../states/contract-state'
+import BalancesState from '../../../states/balances-state'
+import ActivateContractContainer, { calcMaxBlocksForContract } from '../ActivateContract'
 
 
 describe('calcMaxBlocksForContract', () => {
