+++ app/states/__tests__/tx-history-state.spec.js
@@ -1,0 +1,5 @@
+import TxHistoryState from '../tx-history-state'
+
+test('its true', () => {
+  expect(true).toBe(true)
+})
