+++ test/jestSetup.js
@@ -1,6 +1,8 @@
 import Enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
+
 import '../app/fontawesome'
+import './mockDb'
 
 Enzyme.configure({ adapter: new Adapter() })
